Refactor all predictoor agents into one class, PredictoorAgent. All predictoor tests pass
