--- pdr_backend/predictoor/approach1/test/conftest.py
@@ -1,1 +1,0 @@
-from pdr_backend.conftest_ganache import *  # pylint: disable=wildcard-import
