+++ pdr_backend/predictoor/test/mockutil.py
@@ -1,12 +1,8 @@
-"""
-This file exposes run_agent_test()
-which is used by test_predictoor_agent{1,3}.py
-"""
-
 import os
 from unittest.mock import Mock
 
 from enforce_typing import enforce_types
+import pytest
 
 from pdr_backend.ppss.ppss import mock_feed_ppss, mock_ppss
 from pdr_backend.ppss.web3_pp import (
@@ -23,7 +19,7 @@ INIT_BLOCK_NUMBER = 13
 
 
 @enforce_types
-def get_agent_1feed(tmpdir: str, monkeypatch, predictoor_agent_class):
+def mock_ppss_1feed(approach: int, tmpdir: str, monkeypatch):
     """
     @description
       Initialize the agent, and return it along with related info
@@ -32,15 +28,19 @@ def get_agent_1feed(tmpdir: str, monkeypatch, predictoor_agent_class):
     @return
       feed -- SubgraphFeed, eg for binance BTC/USDT 5m
       ppss -- PPSS
-      agent -- PredictoorAgent{1,3,..}
       pdr_contract -- PredictoorContract corresponding to the feed
     """
+    # mock ppss, feed
     monkeypatch.setenv("PRIVATE_KEY", PRIV_KEY)
     feed, ppss = mock_feed_ppss(
         "5m", "binanceus", "BTC/USDT", network="development", tmpdir=tmpdir
     )
+    ppss.predictoor_ss.set_approach(approach)
+
+    # mock ppss.web3_pp.query_feed_contracts()
     inplace_mock_query_feed_contracts(ppss.web3_pp, feed)
 
+    # mock w3, pdr contract
     pdr_contract = inplace_mock_w3_and_contract_with_tracking(
         ppss.web3_pp,
         INIT_TIMESTAMP,
@@ -50,16 +50,11 @@ def get_agent_1feed(tmpdir: str, monkeypatch, predictoor_agent_class):
         monkeypatch,
     )
 
-    # now we're done the mocking, time for the real work!!
-
-    # real work: initialize
-    agent = predictoor_agent_class(ppss)
-
-    return (feed, ppss, agent, pdr_contract)
+    return (feed, ppss, pdr_contract)
 
 
 @enforce_types
-def get_agent_2feeds(tmpdir: str, monkeypatch, predictoor_agent_class):
+def mock_ppss_2feeds(approach: int, tmpdir: str, monkeypatch):
     """
     @description
       Initialize the agent, and return it along with related info
@@ -68,22 +63,25 @@ def get_agent_2feeds(tmpdir: str, monkeypatch, predictoor_agent_class):
     @return
       feeds -- list of SubgraphFeed, eg for binance {BTC,ETH}/USDT 5m
       ppss -- PPSS
-      agent -- PredictoorAgent{1,3,..}
     """
     monkeypatch.setenv("PRIVATE_KEY", PRIV_KEY)
 
+    # mock ppss, feeds
     exchange, timescale, quote = "binanceus", "5m", "USDT"
     coins = ["BTC", "ETH"]
-    feeds = [mock_feed(timescale, exchange, f"{c}/{quote}") for c in coins]
+    feeds: List[SubgraphFeed] = \
+        [mock_feed(timescale, exchange, f"{c}/{quote}") for c in coins]
     ppss = mock_ppss(
         [f"{exchange} {c}/{quote} c {timescale}" for c in coins],
         network="development",
         tmpdir=tmpdir,
     )
+    ppss.predictoor_ss.set_approach(approach)
 
-    ppss.web3_pp.query_feed_contracts = Mock()
+    # mock ppss.web3_pp.query_feed_contracts()        
+    ppss.web3_pp.query_feed_contracts = Mock()        
     ppss.web3_pp.query_feed_contracts.return_value = {
-        feed.address: feed for feed in feeds
+        feed.address : feed for feed in feeds
     }
 
     # mock w3
@@ -100,51 +98,4 @@ def get_agent_2feeds(tmpdir: str, monkeypatch, predictoor_agent_class):
         contract_func,
     )
 
-    # now we're done the mocking, time for the real work!!
-
-    # real work: initialize
-    agent = predictoor_agent_class(ppss)
-
-    return (feeds, ppss, agent)
-
-
-@enforce_types
-def run_agent_test(tmpdir: str, monkeypatch, predictoor_agent_class):
-    """
-    @description
-        Run the agent for a while, and then do some basic sanity checks.
-
-        Uses get_agent_1feed (not 2feeds)
-    """
-    _, ppss, agent, _mock_pdr_contract = get_agent_1feed(
-        tmpdir, monkeypatch, predictoor_agent_class
-    )
-    # now we're done the mocking, time for the real work!!
-
-    # real work: main iterations
-    for _ in range(500):
-        agent.take_step()
-
-    # log some final results for debubbing / inspection
-    mock_w3 = ppss.web3_pp.web3_config.w3
-    print("\n" + "/" * 160)
-    print("Done iterations")
-    print(
-        f"init block_number = {INIT_BLOCK_NUMBER}"
-        f", final = {mock_w3.eth.block_number}"
-    )
-    print()
-    print(f"init timestamp = {INIT_TIMESTAMP}, final = {mock_w3.eth.timestamp}")
-    print(f"all timestamps seen = {mock_w3.eth._timestamps_seen}")
-    print()
-    print(
-        "unique prediction_slots = "
-        f"{sorted(set(_mock_pdr_contract._prediction_slots))}"
-    )
-    print(f"all prediction_slots = {_mock_pdr_contract._prediction_slots}")
-
-    # relatively basic sanity tests
-    assert _mock_pdr_contract._prediction_slots
-    assert (mock_w3.eth.timestamp + 2 * ppss.predictoor_ss.timeframe_s) >= max(
-        _mock_pdr_contract._prediction_slots
-    )
+    return (feeds, ppss)
